+++ inventory/testinventory/serialize/devices.go
@@ -20,8 +20,8 @@ const (
 	bioshake96welladaptorheight                 = 4.5
 	bioshakestandardadaptorheight               = 5.0
 	appliedbiosystemsmagbeadbaseheight          = 12.0 //height of just plate base, upon which most skirted plates can rest
-	appliedbiosystemsmagbeadtotalheight         = 17.0 //height of base and well, in which other plates can rest
-	fluidXhighProfileRackHeight                 = 2.0 - MinimumZHeightPermissableForLVPipetMax
+	//appliedbiosystemsmagbeadtotalheight         = 17.0 //height of base and well, in which other plates can rest
+	fluidXhighProfileRackHeight = 2.0 - MinimumZHeightPermissableForLVPipetMax
 )
 
 const (
@@ -30,7 +30,7 @@ const (
 
 var (
 	incubatorheightinmm = devices.Shaker["3000 T-elm"]["Height"]*1000 + incubatoroffset
-	inhecoincubatorinmm = devices.Shaker["InhecoStaticOnDeck"]["Height"] * 1000
+	//inhecoincubatorinmm = devices.Shaker["InhecoStaticOnDeck"]["Height"] * 1000
 )
 
 // defaultDevices are default devices upon which an sbs format plate may be placed
